Q version 56.81: FR saves and restores mode settings
Not saving / restoring fmode was an accidental omission.

Updated:

 q/fmode.h: Make fmode explicitly 32 bits (rather than ulong).
            When fmode needs to go to 64 bits, will need to use UINT64_C macro
            (from <stdint.h>)

 q/q.c: FR saves / restores fmode

 q/q_version.h: 56.81

 q/scrdit.c: Specify "%o" (not "%lo") when printing fmode.
             When fmode needs to go to 64 bits, should use PRIo64 as fmt
             (from <inttypes.h>)

+++ q/fmode.h
@@ -32,8 +32,8 @@
 
 /* External variables */
 
-extern unsigned long fmode;
-extern unsigned long zmode;        /* Copy of fmode by scmnrd */
+extern uint32_t fmode;
+extern uint32_t zmode;             /* Copy of fmode by scmnrd */
 extern uint8_t xlatable[256];
 extern bool zmode_valid;
 extern int tbstat;
